+++ subler/subler.py
@@ -1,16 +1,16 @@
+# -*- coding: utf-8 -*-
 """This module provides an easily scriptable interface to tagging an assortment
 of media types including x264 video, AAC audio, and many others with iTunes
 formatted metadata via the SublerCLI. By simply creating new metadata
-:class:`Atom`'s and specifying any additionally desired variables in an instance
-of :class:`Subler` you can quickly execute the tagging of metadata to a
-specified file.
+:class:`Atom`'s and specifying any additionally desired variables in an
+instance of :class:`Subler` you can quickly execute the tagging of metadata to
+a specified file.
 """
 import os
 import logging
-import subprocess
 from collections import namedtuple
 
-from .utils import subler_executable
+from .utils import subler_executable, get_output
 
 __author__ = 'Jon Nappi'
 __all__ = ['Atom', 'Subler']
@@ -28,33 +28,36 @@ class Atom(_Atom):
                    'Tempo', 'TV Show', 'TV Episode #', 'TV Network',
                    'TV Episode ID', 'TV Season', 'Description',
                    'Long Description', 'Series Description', 'HD Video',
-                   'Rating Annotation', 'Studio', 'Cast', 'Director', 'Gapless',
-                   'Codirector', 'Producers', 'Screenwriters', 'Lyrics',
-                   'Copyright', 'Encoding Tool', 'Encoded By', 'Keywords',
-                   'Category', 'contentID', 'artistID', 'playlistID', 'genreID',
-                   'composerID', 'XID', 'iTunes Account', 'iTunes Account Type',
-                   'iTunes Country', 'Track Sub-Title', 'Song Description',
-                   'Art Director', 'Arranger', 'Lyricist', 'Acknowledgement',
-                   'Conductor', 'Linear Notes', 'Record Company',
-                   'Original Artist', 'Phonogram Rights', 'Producer',
-                   'Performer', 'Publisher', 'Sound Engineer', 'Soloist',
-                   'Credits', 'Thanks', 'Online Extras', 'Executive Producer',
-                   'Sort Name', 'Sort Artist', 'Sort Album Artist',
-                   'Sort Album', 'Sort Composer', 'Sort TV Show', 'Artwork',
-                   'Name', 'Rating', 'Media Kind')
+                   'Rating Annotation', 'Studio', 'Cast', 'Director',
+                   'Gapless', 'Codirector', 'Producers', 'Screenwriters',
+                   'Lyrics', 'Copyright', 'Encoding Tool', 'Encoded By',
+                   'Keywords', 'Category', 'contentID', 'artistID',
+                   'playlistID', 'genreID', 'composerID', 'XID',
+                   'iTunes Account', 'iTunes Account Type', 'iTunes Country',
+                   'Track Sub-Title', 'Song Description', 'Art Director',
+                   'Arranger', 'Lyricist', 'Acknowledgement', 'Conductor',
+                   'Linear Notes', 'Record Company', 'Original Artist',
+                   'Phonogram Rights', 'Producer', 'Performer', 'Publisher',
+                   'Sound Engineer', 'Soloist', 'Credits', 'Thanks',
+                   'Online Extras', 'Executive Producer', 'Sort Name',
+                   'Sort Artist', 'Sort Album Artist', 'Sort Album',
+                   'Sort Composer', 'Sort TV Show', 'Artwork', 'Name',
+                   'Rating', 'Media Kind')
 
     def is_valid(self):
-        """Performs a check to see if the data in this :class:`Atom` is valid"""
+        """Check that the data in this :class:`Atom` is valid"""
         return self.tag in self._valid_tags
 
     @property
     def data(self):
-        """Return the Subler argument formatted version of this :class:`Atom`"""
-        return '"{%s:%s}"' % (self.tag, self.value)
+        """The Subler argument formatted version of this :class:`Atom`"""
+        return '{%s:%s}' % (self.tag, self.value)
 
 
 class Subler(object):
-    """An OO interface to the SublerCLI"""
+    """A Python interface to the SublerCLI that can be used to easily read
+    from and write to a specified source file.
+    """
     __executable = subler_executable()
 
     def __init__(self, source, dest=None, chapters=None, delay=None,
@@ -75,8 +78,8 @@ class Subler(object):
         :param remove: Boolean flag for remove all existing subtitles tracks
         :param optimize: Boolean flag for optimizing the file by moving the
             moov atom at the begining and interleaving the samples
-        :param downmix: downmix audio (mono, stereo, dolby, pl2) from the source
-            file
+        :param downmix: downmix audio (mono, stereo, dolby, pl2) from the
+            source file
         :param rating: A valid US, UK, or German content rating
         :param media_kind: The type of media represented by the source file.
             Valid values are Music, Audiobook, Music Video, Movie, TV Show,
@@ -88,8 +91,7 @@ class Subler(object):
         """
         self.source = source
         if dest is None:
-            name, ext = os.path.splitext(self.source)
-            dest = name + ' (1)' + ext
+            dest = self._generate_dest()
         self.dest = dest
         self.chapters = chapters
         self.chapters_preview = chapters_preview
@@ -108,27 +110,39 @@ class Subler(object):
         self.metadata = metadata
         self._logger = logging.getLogger('subler.Subler')
 
+    def _generate_dest(self):
+        """Generate a destination file name for the provided source file. This
+        function will increment a number to append to the file name in order to
+        distinguish it from other copies of the same file name
+
+        :return: The absolute path for the destination file
+        """
+        name, ext = os.path.splitext(self.source)
+        counter = 1
+        dest = name + ' ({})'.format(counter) + ext
+
+        while os.path.exists(dest):
+            counter += 1
+            dest = name + ' ({})'.format(counter) + ext
+        return dest
+
     @property
     def version(self):
         """The current version of your systems SublerCLI package"""
-        cmd = '{} -version'.format(self.__executable)
-        output = subprocess.check_output(cmd, shell=True)
-        return output.decode('UTF-8').strip()
+        cmd = [self.__executable, '-version']
+        return get_output(cmd)
 
     @property
     def tracks(self):
         """A list of tracks found the source file"""
-        cmd = '{} -source {} -listtracks'.format(self.__executable, self.source)
-        output = subprocess.check_output(cmd, shell=True)
-        return output.decode('UTF-8').strip().split('\n')
+        cmd = [self.__executable, '-source', self.source, '-listtracks']
+        return get_output(cmd).split('\n')
 
     @property
     def existing_metadata_raw(self):
         """The metadata currently contained in the source file"""
-        cmd = '{} -source {} -listmetadata'.format(self.__executable,
-                                                   self.source)
-        output = subprocess.check_output(cmd, shell=True)
-        return output.decode('UTF-8').strip().split('\n')
+        cmd = [self.__executable, '-source', self.source, '-listmetadata']
+        return get_output(cmd).split('\n')
 
     @property
     def existing_metadata(self):
@@ -160,8 +174,8 @@ class Subler(object):
         """The content rating of the source file. Valid US content ratings are:
         Not Rated, G, PG, PG-13, R, NC-17, TV-Y, TV-Y7, TV-G, TV-PG, TV-14,
         TV-MA, and Unrated. Valid UK content ratings are: Not Rated, U, Uc, PG,
-        12, 12A, 15, 18, R18, Exempt, Unrated, and Caution. Valid German content
-        ratings are FSK 0, FSK 6, FSK 12, FSK 16, and FSK 18.
+        12, 12A, 15, 18, R18, Exempt, Unrated, and Caution. Valid German
+        content ratings are FSK 0, FSK 6, FSK 12, FSK 16, and FSK 18.
         """
         return self._rating
     @rating.setter
@@ -206,18 +220,18 @@ class Subler(object):
         """Apply the specified metadata to the source file and output it to
         the specified destination file
         """
-        cmd = '{} -source "{}" '.format(self.__executable, self.source)
+        cmd = [self.__executable, '-source', self.source]
 
         if self.dest:
-            cmd += '-dest "{}" '.format(self.dest)
+            cmd += ['-dest', self.dest]
         self.metadata.append(Atom('Media Kind', self.media_kind))
         if self._rating is not None:
             self.metadata.append(Atom('Rating', self.rating))
         if self.explicitness:
             self.metadata.append(Atom('Content Rating', self.explicitness))
         tags = [atom.data for atom in self.metadata if atom.is_valid()]
-        cmd += ' '.join(['-metadata', ''.join(tags)])
+        cmd += ['-metadata', ''.join(tags)]
         if self.optimize:
-            cmd += ' -optimize'
+            cmd += ['-optimize']
         self._logger.debug(cmd)
-        return subprocess.check_output(cmd, shell=True).decode('UTF-8')
+        return get_output(cmd)